Describe the origin story of the repository
This project is started in the aftermath of Booster Conf 2023.

I am typing this commit message the day after the conference. And although the Bergen city is basking in the sunlight and is beckoning my to explore her, I am excited to start a collaboration with Einar.

He is a peer I very much respect and lookup to. Einars views are always sensible, well-thought and insightful and is capable of having an open mind and a gleeful curiosity of the world that surrounds him.