Guess the correct Dunham Tiling angle
Earlier I was convinced that numerical instability caused the tiling not to connect up

Boy, was I wrong. I just created a random triangle, but for a triangle to tile, it should have angles that are integral divisors of pi.

In this commit I eyeballed the correct distance for this isosceles triangle to tile.

Next step is to determine the correct points for a triangle with prescribed angles